# feat: spawn the executables directly if possible