@fdncred 2021-08-30
I'm a little concerned that we're not covering all the executables, at least on Windows.  On Windows the `PATHEXT` holds the extensions that are executable. On my Windows system it says this `PATHEXT=.COM;.EXE;.BAT;.CMD;.VBS;.VBE;.JS;.JSE;.WSF;.WSH;.MSC`.