@aminya 2021-08-30
Yes, this pull request currently changes the behaviour for `.exe` files on Windows. That means there is no middle man for spawning exe files. For the rest of the extensions, it shells out to cmd like before, until we implement special care for them too. For example, we can add a function to spawn PowerShell scripts, CMD scripts, etc.