@fdncred 2021-08-30
We also handle ps1 a special way in plugins too here https://github.com/nushell/nushell/blob/66cedf0b3aa4ff4e9b37313ac7a629f4069acf0a/crates/nu-engine/src/plugin/build_plugin.rs#L16-L41