@aminya 2021-08-30
> It also may be worth noting that nu-engine considers these extensions as executable.
> https://github.com/nushell/nushell/blob/66cedf0b3aa4ff4e9b37313ac7a629f4069acf0a/crates/nu-engine/src/plugin/build_plugin.rs#L135-L158

Because only the native executables are spawned directly, this one isn't used yet. But as I said we can implement special code for each of these executables.