@aminya 2021-08-30
> We also handle ps1 a special way in plugins too here
> 

Great! I don't know the relation of this code to `nu-command`. But these two should come together at some point. For now, let's focus on the native executables.