@fdncred 2021-08-30
I'm fine with taking baby-steps toward a goal. @jntrnr, what are your thoughts?