@aminya 2021-09-01
This is ready to go on my end. I have been actually using my branch instead of the official release, which is broken for multiline strings.